refactor: app.pyをモジュール分割してコード品質を大幅に改善 #37
app.pyを6つのモジュールに分割し、保守性とセキュリティを向上

## Changes

### New Modules
- config.py: 設定管理の集約化
- api/database.py: データベース操作の抽象化層
- api/routes_database.py: データベース関連APIエンドポイント
- api/routes_realtime.py: リアルタイム関連APIエンドポイント
- api/websocket.py: WebSocketイベントハンドラー

### Refactored
- app.py: 1,072行 → 176行 (83.6%削減)
  - Flaskアプリケーション初期化に専念
  - Blueprint登録による責任分離
  - 3つのコアルート + エラーハンドラー

## Improvements

### Code Quality
- SQLiteインポート: 8箇所 → 1箇所(集約)
- データベース接続: 15+箇所 → 1クラス(DatabaseManager)
- 単一責任の原則: 各モジュールが明確な役割を持つ
- 依存性注入パターン: 疎結合な設計

### Security
- パラメータ化クエリ: SQL injection防止
- コンテキストマネージャー: 安全な接続管理
- 統一エラーハンドリング: 機密情報の漏洩防止

### Testing
- 全Pythonファイル: 構文チェックPass
- 主要APIエンドポイント: 動作確認済み
- 後方互換性: 維持

## Test Results
- Application startup: Success
- Database connection: 3,266 messages loaded
- API endpoints: All tested successfully
- Syntax validation: All modules passed
