+++ viewer/api/__init__.py
@@ -1,0 +1,4 @@
+"""
+API モジュール
+データベース、リアルタイム、WebSocket APIを提供
+"""
